Added Some Changes to Several Screens
The rights on several screens works. The Available Inventory Report is now broken. Change History report now has some code in for it, but it doesn't do anything. Added APIs for the search buttons to trigger the stored procedure, but still need more info from David on how to implement it.